# Conflicts:
#	spark-backend/src/buckets/buckets.service.ts
#	spark-backend/src/buckets/entity/habit-bucket.entity.ts
#	spark-backend/src/habits/habits.service.ts
#	spark-backend/src/users/entities/user.entity.ts

+++ spark-backend/src/goals/goals.service.ts
@@ -1,4 +1,4 @@
-import {BadRequestException, Injectable, NotFoundException} from '@nestjs/common';
+import {BadRequestException, ForbiddenException, Injectable, NotFoundException} from '@nestjs/common';
 import {CreateDailyGoalDto, CreateGoalDto, CreateIterativeGoalDto, CreateScheduledGoalDto} from './dto/create-goal.dto';
 import {InjectRepository} from "@nestjs/typeorm";
 import {Repository} from "typeorm";
@@ -6,6 +6,10 @@ import {Habit} from "../habits/entities/habit.entity";
 import {Goal, GoalType} from "./entities/goal.entity";
 import {UpdateGoalDto} from "./dto/update-goal.dto";
 import {GoalDto} from "./dto/goal.dto";
+import {UserPayload} from "../users/auth/user-payload.model";
+import {User} from "../users/entities/user.entity";
+import {FriendBucket} from "../buckets/entity/friend-bucket.entity";
+import {HabitBucket} from "../buckets/entity/habit-bucket.entity";
 
 @Injectable()
 export class GoalsService {
@@ -13,16 +17,27 @@ export class GoalsService {
         @InjectRepository(Goal)
         private goalRepository: Repository<Goal>,
         @InjectRepository(Habit)
-        private habitRepository: Repository<Habit>
+        private habitRepository: Repository<Habit>,
+        @InjectRepository(FriendBucket)
+        private friendBucketRepository: Repository<FriendBucket>,
+        @InjectRepository(HabitBucket)
+        private habitBucketRepository: Repository<HabitBucket>
     ) {
     }
 
-    async create(createGoalDto: CreateGoalDto) {
-        const habit = await this.habitRepository.findOneBy({
-            id: createGoalDto.habitId
+    async create(createGoalDto: CreateGoalDto, currentUser: UserPayload) {
+        const habit = await this.habitRepository.findOne({
+            where: {
+                id: createGoalDto.habitId
+            },
+            relations: ["user"]
         });
         if (habit == null) {
             throw new NotFoundException("Habit not found");
+        }
+
+        if (habit.user.id !== currentUser.id) {
+            throw new ForbiddenException("User in goal must match current user");
         }
 
         let goal: Goal;
@@ -64,26 +79,94 @@ export class GoalsService {
         return GoalDto.fromEntity(await this.goalRepository.save(goal));
     }
 
-    async findAll() {
-        return this.goalRepository.find().then(goals => goals.map(goal => GoalDto.fromEntity(goal)));
+    async findAll(currentUser: UserPayload) {
+        const userGoals = await this.goalRepository.find({
+            where: {
+                habit: {
+                    user: {
+                        id: currentUser.id
+                    }
+                }
+            },
+            relations: ["habit"]
+        })
+
+        const friendGoals = (await this.friendBucketRepository.find({
+            where: {
+                users: {
+                    id: currentUser.id
+                }
+            },
+            relations: ["habits"]
+        })).flatMap(bucket => bucket.habits.flatMap(habit => habit.goals))
+
+        const bucketGoals = (await this.habitBucketRepository.find({
+            where: {
+                users: {
+                    id: currentUser.id
+                }
+            },
+            relations: ["habits"]
+        })).flatMap(bucket => bucket.habits.flatMap(habit => habit.goals))
+
+
+        return userGoals
+            .concat(friendGoals)
+            .concat(bucketGoals)
+            .map(goal => GoalDto.fromEntity(goal));
     }
 
-    async findOne(id: string) {
-        const goal = await this.goalRepository.findOneBy({
-            id: id
+    async findOne(id: string, currentUser: UserPayload) {
+        const goal = await this.goalRepository.findOne({
+            where: {
+                id: id
+            },
+            relations: ["habit.user"]
         });
         if (goal == null) {
             throw new NotFoundException("Goal not found");
         }
+
+        const friendGoalIds = (await this.friendBucketRepository.find({
+            where: {
+                users: {
+                    id: currentUser.id
+                }
+            },
+            relations: ["habits"]
+        })).flatMap(bucket => bucket.habits.flatMap(habit => habit.goals.map(goal => goal.id)));
+
+        const bucketGoalIds = (await this.habitBucketRepository.find({
+            where: {
+                users: {
+                    id: currentUser.id
+                }
+            },
+            relations: ["habits"]
+        })).flatMap(bucket => bucket.habits.flatMap(habit => habit.goals.map(goal => goal.id)))
+
+        if (goal.habit.user.id !== currentUser.id ||
+            goal.id in friendGoalIds ||
+            goal.id in bucketGoalIds) {
+            throw new ForbiddenException("Cannot view goal");
+        }
+
         return GoalDto.fromEntity(goal);
     }
 
-    async update(id: string, updateGoalDto: UpdateGoalDto) {
-        const goal = await this.goalRepository.findOneBy({
-            id: id
+    async update(id: string, updateGoalDto: UpdateGoalDto, currentUser: UserPayload) {
+        const goal = await this.goalRepository.findOne({
+            where: {
+                id: id
+            },
+            relations: ["habit.user"]
         });
         if (goal == null) {
             throw new NotFoundException("Goal not found");
+        }
+
+        if (goal.habit.user.id !== currentUser.id) {
+            throw new ForbiddenException("Can only update current user's goals");
         }
 
         Object.assign(goal, updateGoalDto);
@@ -101,12 +184,19 @@ export class GoalsService {
         return GoalDto.fromEntity(await this.goalRepository.save(goal));
     }
 
-    async remove(id: string) {
-        const goal = await this.goalRepository.findOneBy({
-            id: id
+    async remove(id: string, currentUser: UserPayload) {
+        const goal = await this.goalRepository.findOne({
+            where: {
+                id: id
+            },
+            relations: ["habit.user"]
         });
         if (goal == null) {
             throw new NotFoundException("Goal not found");
+        }
+
+        if (goal.habit.user.id !== currentUser.id) {
+            throw new ForbiddenException("Can only remove goals of current user");
         }
 
         await this.goalRepository.remove(goal)
